requirements.txt
APScheduler==3.9.1
google_api_python_client==2.39.0
google_auth_oauthlib==0.4.6
pandas==2.1.2
protobuf==3.19.4
python-dotenv==0.19.2
twilio==7.3.2
Unidecode==1.3.2
pytest==7.1.3
names==0.3.0